examples/python/requirements.txt
pyzmq==14.4.1
